refactor: separate youtube module
