+++ src/app/app.routes.ts
@@ -3,7 +3,7 @@ import { Routes } from '@angular/router';
 export const routes: Routes = [
   {
     path: 'search',
-    loadComponent: () => import('./core/pages/search.page').then((m) => m.SearchPage),
+    loadComponent: () => import('./youtube/pages/search.page').then((m) => m.SearchPage),
   },
   {
     path: '',
