--- src/app/core/models/video.model.ts
@@ -1,3 +1,0 @@
-import { VideoDto } from '@/shared/services/api.dto';
-
-export interface VideoInterface extends VideoDto {}
