+++ src/app/youtube/components/item/item.component.html
