+++ src/app/youtube/models/video.model.ts
@@ -1,0 +1,3 @@
+import { VideoDto } from '@/youtube/services/api.dto';
+
+export interface VideoInterface extends VideoDto {}
